requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Babel==4.0.0
Flask-Mail==0.10.0
PyMySQL==1.1.1
bcrypt==4.1.3
pdfplumber==0.11.1
Werkzeug==3.0.3
WTForms==3.1.2
cryptography==42.0.8
python-dotenv==1.0.1
Pillow==10.3.0
